fix(cb2-3462): reconfigure project so handler.ts only exports handler
