+++ src/Interfaces/INewTechRecord.ts
@@ -1,0 +1,3 @@
+export interface NewKeyStructure {
+  [index: string]: string | boolean | number | Array<string>;
+}
